+++ app/Http/Controllers/Databases/CreateDatabase.php
@@ -6,26 +6,25 @@ use Exception;
 use Illuminate\Http\JsonResponse;
 use Servidor\Database;
 use Servidor\Http\Controllers\Controller;
+use Servidor\Http\Requests\Databases\NewDatabase;
 
 class CreateDatabase extends Controller
 {
-    public function __invoke(Request $request): JsonResponse
+    public function __invoke(NewDatabase $request): JsonResponse
     {
-        $data = $request->validate([
-            'database' => 'required|string',
-        ]);
-
-        $db = new Database();
-        $dbname = (string) $data['database'];
+        $database = $request->validated();
 
         try {
-            $created = $db->create($dbname);
+            $created = (new Database())->create($database['name']);
         } catch (Exception $_) {
             $created = false;
         }
 
         return $created
-            ? new JsonResponse(['name' => $dbname])
-            : new JsonResponse(['error' => 'Could not create database'], 500);
+            ? new JsonResponse($database)
+            : new JsonResponse(
+                ['error' => 'Could not create database'],
+                JsonResponse::HTTP_INTERNAL_SERVER_ERROR,
+            );
     }
 }